<refactor>: LNB + RNB => MenuBar
